`serialize-impl` feature on `diffus-derive`.
Previously, `diffus-derive` generated
`#[cfg_attr(feature = "serialize-impl", derive(serde::Serialize))]`
which caused the evaluation of the expression to be performed when
`#[derive(Diffable)]` was evaluated, i.e. in the end-users project.

This change introduces the `serialize-impl` feature in `diffus-derive`
so that a `#[derive(serde::Serialize)]` is put on `Edited*` types
whenever the end-user is using `#[derive(Diffable)]`.